OK, so Linnormal is still fastest.
At least, when there isn't some freaky slowdown from the RNG, and that's hard to predict... It seems to happen to various generation methods. I've benchmarked all of the ones from before with GoldenQuasiRandom instead of PouchRandom, and the results seem more consistent. It's hard to be sure; using GoldenQuasiRandom makes the generate-next-long step take almost no time, which would be quite different with java.util.Random .