Fury is also very fast at reading.
Tantrum doesn't seem to help, but at least doesn't hurt much? It may be possible to speed it up... It also may be possible to store the ordered maps and sets simply as key-value pairs in order, and not explicitly store the list for order. I don't know if Fury does that now.